Update ObjectSelectorWindow.cs
Your package already includes the .meta files which you can use them to find without iterating through every directory. The Unity is smart enough to know if there is an asset which has the same GUID and gives an error on the console. But that is nearly impossible because a GUID is too big.